med-agenda-complete/med-agenda-project/backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dateutil
aiofiles

pytest
httpx
pytest-asyncio

python-dotenv
aiosmtplib
twilio
pytest-cov
coverage
httpx
